Add files via upload
Issue was that text contained multiple whitespaces together and in this file when text.split is done all the whitespaces were being eliminated. This was causing discrepancies in index count. So modified split function to not delete extra whitespaces. This is working fine now. Result can be seen in output1.txt. Also, input and output path is hardcoded so modify when you are running.